Move sinkpit_scheduler in schedulers (HUB)
